refactor(#5): split sharing library into create/manage/queries modules
Keep sharing.ts as a barrel re-export so existing imports stay stable.
